Objects: Moving objects with physics
This is mostly for testing the ODE solvers and seeing that we have a working scene.
Right now the objects are trying to be a distance of 15 from every other object.
This obviously is not possible for every object to be exactly 15 from the others,
but it will optimize. Each mass can be a different size and have a different mass,
causing it to behave differently.